Better checking for exceptional issues...
Better checkpoint progress testing so that I can stop the processes at
it's earliest indication of failure to prevent clobbering good files
with a bad one generated from faulty process.  Also doing a quick test
for the root key for the keystore to know if I should remove it or
not.