Insulating the filesystem from DRS
Created a filter to get URLS that are sent in by the client that
follow the DRS syntax.  This is the scafolding for querying the
database to provide the resolution to the actual filesystem file.
There is support for url query type strings as well as the expected
straight up path urls.

Next: get the query in here, test the resolution and then connect the
filter.

Hmmm... there may be two looming issues.
1) Database committing properly
2) Threading issues as I am not sure I have locked that down yet.
TBD