Changing how keys are generated and installed - using hybrid openssl / keytool mechanisms
Will not use openssl for initial setup.  Will continue to use keytool
and generate the initial keystore and setup the installation.  Then
will use openssl to generate key and csr.  When the csr is signed and
returned as a cert will use openssl along with keytool to install into
the system.

TODO: Test... and come up with a single way of describing and feeding
the DN to keytool and openssl so that it does not have to be done
multiple times by the user.