Tomcat configureation and Key and Certificate juggling...
Put the configureation of tomcat in a separate function from the
installation of tomcat, though right now I call the configuration
function from inside of the setup_tomcat install function.

When generating ssl key and csr I have now put in a placeholder DN
value for the generated cert request to use.  In the future I would
like to first check if there is a java keystore, then pull out the dn
from that and refashion it to be used for the dn for the ssl generated
cert.  This is in effort to manage the complexity of having multiple
keys and DNs etc.

This is a flux commit since at this moment I have not had a chance to
fully test it.

:-)