Maintain DNs between tomcat's keystore and globus' grid-security key
This extends the basic idea which is the minimize the skew between the
Tomcat keys, in the keystore, and the globus keys, in
/etc/grid-security.  The impetus for having these two formats still
represent the same underlying key (which includes same DN
ofcourse). Is simply to minimize maintainance issues that may arise.
For example keeping up with expriation dates and which CAs signed
which key, etc.  Perhaps this is not the way to go, time will tell.
For now I would like to er on simplicity in the key managment model,
though that makes things a bit more complicated in the script.

:-|

This commit has a new helper function to extract the DN from the
keystore for use by the openssl generated key pair needed for globus.
The calls to "generate_ssl_key_and_csr" function now also takes the
keystore password used in the new "extract_keystore_dn" function.
Sensible defaults have been made avaialble and variable scoping in
these functions have been tightened. (This scoping helps minimize the
exposure of the password variable, somewhat)