Aync_with_java_truststore function and using modifed InstallCert class
Everytime we want to manipulate the truststore we should sync that
truststore with the java jre truststore (jssecacerts) file.  The bug
that shoudl be resolved in this commit... I know I know... this commit
may be a bit premature.... However, the problem was that InstallCert
was reading in the java jre jssecacerts file since it didn't find one
locally.  The reason why it didn't find one locally is because it was
explicitly looking for a local file named jssecacerts, else it would
use the jre's jssecacerts or it would use the jre's cacerts file as
the base keystore in which to add this new certificate.

So I modified the InstallCert code to take an additional arg where you
specify the base keystore file that you want it to add to.  So now I
can point to a file named something other than jssecacerts, for
example esg-truststore.ts.  This should address why the local key was
not being seen in the final keystore, but the remote server's cert was
there (from the call to register).  Things should be all good now.
Let's cross our fingers :-)