Made check_tomcat_process smarter...
Now it will scan tomcat's server.xml file for defined ports and then
check those ports to see if they are active.  It also checks the
process names bound to those ports.  If the names are not all jsvc the
user is warned and shown the process names of the interloping
process(es).

Still returns a non-zero value under any condition that is not
'running'.

return values....

No process running: 1 or 2* ->  none found
another process bound to port: 3