Made setup_esgcet() a bit smarter w.r.t. upgrade vs installation added check_module_version
So now the setup_esgcet uses the new check_module_version function to
figure out if it needs to be upgraded or installed.  The result of
this detection is used downstream to set the choice prompts
appropriately for the user - so they can just hit enter. As well as
using that "mode" indicator to select if there needs a reconfiguration
needs to be done or not.  Overall a smarter setup function for esgcet.

The lengths I'll go for my users ;-).