Postgress access...
So created a file to hold postgress password.  The password is read
out of that file if it exists and is used throught the script when
database access is required.  The password file is put in the home of
the installation user.  The value is written out to file after
successful postgres installation (and read back and used
subsequently).
The value can be overridden if pg_sys_acct_passwd is set in the
environment that launches the esg-node script.