Switches work... writing config type to {esg_root_dir}/config/config_type
So, start must be the LAST arg read therefore --type data &| index
must be before it or they won't get read!  The --type args set the
$sel var that is filtered throug bit vectors that tell the script
which functions to run to achieve a particular node configuration
type.  As of this commit there are only two data and index (for data
node and index node respectively).

When a successful start is done it writes out the $sel (--type)
numeral value to file.  Subsequently start may be the only argument to
esg-node because it will read the needed $sel value from the written
{esg_root_dir}/config/config_type.

I have put this bit vector filtering also as a part of the installation.

NOTE: It is only in the case of start stop status restart where the
--type flag and arg(s) must precede.

Also....
Set default values for question on "Do you want to continue wth FOO
installation..."  to [Y/n], yes.  This has been modified where ever a
check_*_version call can be made.

:-\ I think this should really be nearing a close on the scripting!
-gavin