Put the reading of config_type closest to where it would be used... start()
it is the start function that really needs to know "what to start?" so
if no selection as been made to that point then in the start function
we will read the config_type file and set the value appropriately.
At the end of the function, the value is written out.  It represents
the last successful start --type configuration.

Remember the very first call to start (where there is no config_type
file) requires the --type to be specified in the preceding arguments.