Created test_globus() function and set it and setup_globus to take and decipher type selection.
Dealing with the globus setup is "interesting".  I wrote it based on
the notion of sequestering all globus tool machinations into a single
location.  It serves two configurations as it was the same file used
for "gateway" and "data node" installations.  Now, there is no longer
a separate "gateway" installation script. But the esg-node script now
has the notion of different "types" of node installations, so I am
calling the already dual-purpose globus script with the appropriate
args for the appropriate type.