New switches in place for data/index/idp nod configurations
This commit has a lot of different changes in it.  The primary change
is having working BIT variables to allow us to select the
configuration type for the node installation.  Because of the dual
nature of the globus script calls to it are now qualified using the
bit fields.  The nomenclature of "gateway" is not present, except for
its use when calling globus, which still has the concept of "gateway".