Added commons lang and httpclient to dependencies...
This is to support the use of ant's publishing mechanism in hopes that
we can publish to the artifactory repository.

see:
http://ant.apache.org/ivy/history/latest-milestone/use/settings.html
"HTTP Authentication
Note: HTTP Authentication can be used only if commons-httpclient.jar
is in your classpath If any of the url you use in ivy (especially in
dependency resolvers) need http authentication, then you have to
provide the host, realm, username and passwd attributes of the
configure task. These settings will then be used in any further call
to ivy tasks."