Enhanced with better arg negotiation and smarter use of type directives and update logic.
Lots of changes here.  This fixes some issues with the move to node
"type" base installation.  There are some enhancements to give
"smarter" prompting, and more argument checks and helpful feedback to
the user.

There is still an issue with postgress installation.  The creation of
a database user account.  This opens up the machine for login using
the database system user.

This is a valid departure from the master branch with the advent of
these new features.