tailoring make_fresh_keystore for use inside of esg-node
Checking for download 404s by inspecting recieved content.
I am not fully comfortable with this because it is not a complete
solution.  There may be other return types that are equally
"unsuccessful" that won't be caught. :-\

In anticipation of moving the tomcat configuration into the /esg
directory structure, created tomcat_conf_dir variable where necessary.
(not changing anything related to the location of server.xml though)