Use <module>-web.xml file from build dir not from conf...
The <module>-web.xml file is copied from conf to build directories
during the build process.  Then when a war is being created the
<module>-web.xml is used for the war's web.xml file.  The key change
here is that the war uses the file from the build directory.  This
gives us the opportunity to do some token filtering during the copy
from conf to build before it gets used in the war.  Also we want to
exclude there being two copies of the web.xml files in different
locations with different names, so we exclude it from the war's
classes directory, where all the other property files go.