2 Mods: added --myproxy flag and modified where passwords are stored
passing through args specifically for use for myproxy (gateway/idp
configuration).  Much like the --gridftp-config flag we can pass
through args to myproxy as well.

it is not wise to write down passwords in an open file of environment
variables :-)... so... what I am doing now is writting down the
password to the database and keystore into the esg root in dot files
chmod'ed 600.