Prep'ing to move tomcat configuration elsewhere
It would be best to move the tomcat configuration files (keystore,
et. al.) elsewhere so that even if tomcat goes away users won't have
to screw with keystores and such.

fixed the dispatch to the globus method: Explicitly setting specific
byte to get the desired path to be taken.

re-organized verbiage (usage) so things are more clear.

I have the tomcat config directory pointing to under tomcat, when I
tried it last it didn't work quite right... :-(  I'll have to debug
what's the deal but hey... at least it is prepped.

A few good cleanups and setting of better, more explicit, full path
default values.