fixed bug recently introduced dealing with setup_globus() false error return
The scoping of ret was wrong and and thus the checked_done call at the
end was returning a false negative.  So now properly returning 0
explicitly from the setup_globus() function.  (the ret value setting
and resetting is a vestage from the days when I would pass in a
selection value that was a combination of both handled types.  The
function would handle them in a single call.  I have since resorted in
calling the function specifically with either one or the other type
singularly, explicitly set so that each invocation of the setup_globus
call would take a single execution path.)

Yes, I could make separate functions, but there is not a strong enough
reason to separate these strongly tied sematic reasons. IMHO.