not posting war or jar of full project
The problem is that I can't figure out how when I add the descriptor
to the ivy.xml file how to specify on the downloading side which jars
to get.  Right now the calling side gets all or nothing of what gets
posted.  So I decide to post fewer things.  Not the best solution but
right now it will do.

I can see not posting the war.  But it would be nice to be able to
post the full jar but not have everyone who comes getting everything.

This is on the TODO list.