Fix bug in database creation script calling - had to re-adjust the max type value
I would find myself in an inifinit loop if i did
$> esg-node --type ALL --config-db

That is because the configure_postress() function that gets called
decrements each of the types each time around the loop.  Thus If have
a larger value than I have values to decrement then I never get to
zero, the condition that stops the loop.  Now I make sure that the max
is set only to the value equal to the sum of the flags I am able to
decrement.