Created object id for UserInfo based on tuple of fistname+lastName+openid
This object represents a user in the system, encapsulating their basic
information.  By using this tuple it will be easier to see if users
are equal at the object level. Because there are other attributes that
do not take part in the equality, writers most be extra careful when
using data structures to hold these objects that they be sure to
remove the object and write put in this new one during insertion.
Many datastructures will refuse to re-write an object that is
considered already there!