Essentially reverting last changes to UserInfo.java and UserInfoDAO.java
The changes I made were in attempt to provide some code level security
to the user info object.  Also attempted to do something smart with
the equals contract.  This is something to review.  The basic question
is what does it mean to be equal and what things are immutable about
a user info object?

I'll have to get back to this, but for now I am rolling back a bit to
keep things simple and deffer this decision until after everything
else is setup and tested.