Created Data objects and Data Access objects to support IDP functionality
Still have work to do... right now the query for creating the UserInfo
object from  the UserInfoDAO is not fully populating the object.  Also
there needs to be a decision about the openid.  What is the best way
to set the base of the openid? What is the best way to create it?
When should it be created?

And I still have to test some more.