Unit tests for Event Queuing
Event Queues work remarkably well, at least the annecdotal evidence is
promising.  The system events are good and work.  There may be a small
degree of coordination with respect to the order of event propagation
in a broadcase scenario.  At the moment using a hashmap so no order
promises... but something tells me that I should probably be
broadcasting in connection order?

Anyway so far so good