Finished Registration Merge and added RegistryUpdateDigest object
The merge of the registration is "complete".  The result is a new
node list for our local registration object and the set of all the nodes that have
been modified or added as a result of this merge.  The
RegistryUpdateDigest is simply a holder for the payload xml document
string that is the registration.xml text, the checksum value for that
text and the set of updated nodes.

This is passed down the event queue chute, where the next component
(the connection manager) will take it and peform the second half of
the gossip protocol i.e. selecting two random nodes to send our new
state to as well as feeding back to the ESGFRegistry those nodes that
have timed out.