Registry is now able to respond to peer nodes leaving...
When peer nodes leave, a "join" event is generated that the Registry,
as all components, gets to respond to.  specifically the event when
nodes leave.  This now triggers the removal of that node object from
the gleaner's internal JAXB XML representation using the new map we
created in the gleaner. (see that commit for more details).