Now Has ability to remove Nodes from structure.
In order to be able to remove nodes from the internal xml object
structure created by JAXB, I had to create an additional look up map
[hostname -> node].  This map is "sync'ed" via the sync() method
called in a few places where we want to make sure... both
representations stay in... sync!

This is particularly important in the case when the registry adds
nodes.  If the internal map doesn't know about the additional adds
then when a remove event comes along it won't know about the node to
remove it.