Connection Manager knows to add newly discovered peers
To add a new peer, the connection manager sends a message to the node
manager to "register" the new peer.  Then connection manager is in-turn
sent a JOIN event from the manager that the node has joined the system
and then the connection manager behaves accordinly to add the node to
it's data structure(s).

Also now there are two types of QueuedEvent handling... one for
getting registry updated of its world view.  The other is to pass
through remote events.  This needs more looking into... but the idea
is that nodes should be able to route events from node to node.
Think about the gossip search event propagation scenario.