Make P2P Gossip more active...
The way the protocol worked was that disseminating information to
peers is fine.  However, at the momemt peers are pinged explicitly by
each other, which is fine - though could be refined to be more of a
leasing scenario (for later).  The problem comes in when for some
reason peers cannot ping each other and purges each other from their "world
view".  So far so good... but how do then they re-connect if the
system has already reached quiescence?

Well, as it was written, they don't and thus network failures post
quiescence means that there is no "mesh repair" possible.  So to fix
this each node will "touch" its registration and send it out
periodically to stir things up a bit and let everyone know that there
it is still there and it's information is upto date.

The frequency of this is determined by the property:
"registry.period" which is now defaulted to 30 minutes (will tweak for
best fidelity).

-gavin