Accomodates not having passwords written to propety files
Instead of having passwords written to the configuration property file
and then having to lock down that file because it has passwords in
it... I have modified the installaiton script esg-node such that it
puts passwords in hidden files that are chmod'ed appropriately.

With that in place now this ESGFProperties object had to be smart
enough to fetch the data out of these files and incorporate them into
the property representation as seemlessly as possible.

Right now there is support for reading the overall password and the
database password files' info.

-gavin