Added two simple "built-ins" to set and unset env also...
Added "exit" logic to first reset your identity back to base user and
then exit out of shell completely from there.

Added "su", "id" and "env" built-ins.
The su (though currently shorted for testing) sets the appropriate
environment settings for root user "rootAdmin".  You must be running
the shell local on the node to run as rootAdmin (by design).  Though I
see making this shell be able to establish remote connections I
believe that root should be only when on the local machine. (but as
long as we keep things stream i/o based all of this is gravy.

Also, setting the prompt accordingly based on environment settings.

The notion of environment that is deeply exercised in this code
facilitates a more functional programming style where data structures
encapsulate state and functions are "pure" with little to no side
effects.  This makes this thread safe (though not a huge deal in a
shell context), and keeps necessary shared data in one place.

The set and unset builtins are implemented as inner classes.
The "test" and "clear" provide examples for how to build stand along
command classes.