Cleaned up framework for Shell and shell command mechanics
Commands are created and their instance hang around for the duration
of the shell.  There was some sticky state with the command line and
options that was causing some strange behavior when OptionGroups were
used.  Essentially, the option value would not get reset after an
evaluation. Thus subsequent evaluations would conflict with the state
of the options set previously.  Since apache CLI was not intended to
exist beyond a particular call it does not make an effort to clean up
after itself since it was intended to exit after the run was
executed.  In a shell context this is not the case so I had to create
a way to "reset".  Methods were delegated through and lazy
instantiations made where necessary to provide flexibility.

All commands have been updated accordingly.

there is one thing that is curious.  I made a method abstract but it
doesn't seem to like to be called from the constructor.  Maybe this is
an esoteric java rule I failed to memorize... but I doubt it.

Hmmm... curious