Gossip Protocol tweaking....
The issue really is that the protocol up until now was overly
ambitious w.r.t. reaching quiescence.  If a message was seen before by
a node, the node would drop that message and not allow it to propagate
any further.  Though you reach quiescence relatively quickly it is at
the cost of blocking other nodes that have not gotten that message
(which is new to them) from getting it.  This is not a total problem
because of the randomness introduced in selecting the next hop value.

So, now I am trying to establish the use of the TTL so be the
mechanism to limit message propagation, not just the fact that this
message has been seen before by a node.  This will make the protocol
more noisy but will also make it more responsive to changes in the
federation.

I suspect that there may be a hybrid solution dependent on the type of
messages that are being sent.  For example, I plan to put in a
"goodbye" message that should rattle around the network and be
completely based on TTL and the use of the origin field.  Thus these
goodbye messages will follow more of a "gossip mongering" type of
behavior rather than the registration sync'ing "anti-entropy" type of
behavior.   So this commit is the beginning of this experiment.  But
regardless of behavior this commit starts in a good place

1) The at the ingress (service) level, TTLs are inspected and adjusted
before getting pushed through the FSM.

2) [a bit more dubious/experimental] The Registry component doesn't
drop messages it has seen before.