Fixed ConncurrentModificationException in ConnectionManager...
Exception in thread "Timer-5" java.util.ConcurrentModificationException
	at java.util.HashMap$HashIterator.nextEntry(HashMap.java:793)
	at java.util.HashMap$ValueIterator.next(HashMap.java:822)
	at esg.node.connection.ESGConnectionManager.pingToPeers(ESGConnectionManager.java:162)
	at esg.node.connection.ESGConnectionManager.access$000(ESGConnectionManager.java:99)
	at esg.node.connection.ESGConnectionManager$1.run(ESGConnectionManager.java:156)
	at java.util.TimerThread.mainLoop(Timer.java:512)
	at java.util.TimerThread.run(Timer.java:462)

The problem was that I was iterating through a live datastructure that
was being modified as a consequence of the iteration.  So now I create
a temporary datastructure to hold the info, thus providing a separate
iterator and not having a conflict.

This was making it so that nodes could not purge stale or inactive or
unresponsive peers that were detected.  This is why there are so many
OLD/"BAD" entries in the registries.