Timestamping... only when registrationis modified in some way...
When nodes are removed or nodes are added will trigger a timestamp
written to the registry.  This makes clear the semantics, that the
time on a regigistration entry is the time it was last modified in
some way.  There is no other condition that should result in a time
stamp modification.