Egress portion of transporting remote events through peer network.
The main thing is setting up events to be dispatched appropriately.
Created some helper methods for dealing with common manipulation of
[Remote]Events.

Lots of changes I am a bit too tired to get into right now... but
basically I am routing unregister events through the registry to the
connection manager and then back out to random peers.  On ingress the
service decrements the TTL.  Connection manager checks TTL (via
isValid call on RemoteEvent) if the TTL is down to zero no more
forwarding. This is not just for the UnRegister but also for Register
events.