Active node removal trigger for push
When a node is detected to be gone the peer proxy object that
determines this tells the listening communications manager about this
and then the manager moves the proxy to a secondary datastructure
before trying it again and removing it.

With these modifications, the removal of a node that causes a change
in the registry triggers a dispatch of the new registry state to the
rest of the peer network.  Now there is symmetry with regards to state
dispatches.  It happens when new nodes are added as the result of the
merge of an incoming peer registration document, as well as (now) the
discovery of the removal of a peer.  So anything that changes
registry state is propagated.

Note, the ping discovery mechanism can be turned into a more
sophisticated leasing system as an optimization.

Also, node information that does not result in a change in state are
forwarded from us AS-IS (except for the decriment in TTL) to the next
two randomly selected peers.  Note when a change in state does occur
we send a brand new dispatch event to two randomly selected peers.
The rationale is that our representation is subsumes the incoming
representation that changed us and thus by pushing out state there is
no information lost in the forward progression.