Using new function in esg-node calling script get_current_webapp_version
If the user decides not to upgrade the node manager the version of the
node manager should be set to the current version so that downstream
code that uses the node manager version are using the current version
that is in play.  This is needed for the filters to do the right
thing... i.e. select the right version of the node manager jars to
install.
:-) gavin