Addressed connection manager's timing issue w.r.t periodic tasks
The connection manager periodically will send out it's
registration.xml to it's peers.  There is a situation during startup
where if a preceding operation takes 'too long' (longer than "delay"
time) to start, the registration.xml file that is expected to be there
may not be and thus this cascades into a parsing error and the
exception causes the P2P mechanism to fail.  Now we catch that (or
any) exception that may occur when reading the expected
registration.xml file and simply balk till the next call of the timer
until it is there.

The larger question to re-visit here is to reconcile this mechanism of
the ESGFRegistry already doing this with it's timer.  It seems like
the proper next-step fix is to not call the function here
"periodicallyRegisterToPeers" and let the ESGFRegistry trigger this
from the timer started in ESGFRegistry.startRegistry().

For now this is a band-aid fix... Next: test removing the function in
this class.  It will also have the side effect of limiting network
chatter, which is a good thing! :-)

(made some good comments on how to address issues around
'pingToPeers' and how that can be made tighter)

-gavin