Also putting group (attribute) filtering at genisis of local registration document
The local entry into the registration.xml should not have groups that
are not "shared" / advertised to the rest of the federation.

Groups: User, Guest, Nobody, Root
are always implicitly "private" and not advertised.

These groups are largely vestiges from the Gateway and brought over
via the user migration tool.  The "Root" group should be wholely removed
in lieu of the "wheel" group (perhaps a TODO for the migration code).
The "User" group is lower cased in P2P to "user" and it is not clear to
me that "Nobody" should exist.  The Guest group is fine.  All groups
should not be advertised across the federation.