Provide new boolean directive for doing "full checks (poking)" - for now only during shard file creation...
So to allow for the ability to do PORT LEVEL pruning of services, I am
passing a boolean (doCheck) to indicate this deeper level of
inspection.  The goal is to be able to prune unresponsive services at
the port level.  Before this the pruning was limited to only looking
at the node-manager service availability.

"poking" is what I call trying to tcp connect to a remote port (that a
service is expected to be found on).  Like pinging... but at port
level. Just something I made up ;-).  See Utils.poke(...)

Right now when a peer signals that it has left there is a call to do
poking. (this should go away once I do the next round of changing to
connect pruning to be the trigger for registration rebuilding with
poking)