WIP: Now have routable 'system' events
Set up a new event, "ESGFPruneEvent".  This class is an instance of
ESGCallableRoutableEvent which comes from the lineage of event
classes.  The key contribution here is that these 'routable' events
allow for the creation of events that have their own path through the
stages (components), they don't have to follow the predefined
overarching FSM that defines the connections between states
(components). This new class of events provids greater flexibility.
They are intended to be used for "system" events.

Notes:

Right now, in ESGFPruneEvent I am bastardizing the <T> generics
I set up in ESGFCallableRoutableFutureEvent and parent
ESGCallableFutureEvent.  I am forcing a Boolean type.  For the
specifics of the PruneEvent this is okay, however, this imposes that
the calles made within have to post their "data" as Booleans as well.
Hmmm... I should be more amenable to other things.

The idea is that as the routable events goes from state to state it
can also collect data and work on data that resulted from the visited
stage's transformations.  Typically event passing results in the event
getting changed and all state recorded in the event.  These special
events behave in such a way that they visit stages, but they can
capture data and preserve intra-stage context to perhaps do something
more informed with the data it sees through it's travels.

Essentialy this event can be an aggregate recording of event
manipulations.  It is a cross between an "event" and "command"
object.