Trying to be as clean as I can with memory for routable events
The routable event has a List in it and a Map.
For the list, the goal is to create a list that is just the right
size.  Also made sure there was a check so the call to getNumHops()
won't throw up an NPE.

("next" I'll do something about the map sized) ;-)

Oh also... regarding the tiny RegistrationGleaner change... pulled the
sync() call out, so that it is not called inside of the
saveRegistration() method but along side it.  I am not sure that this
will make any difference whatsoever (probably not), but as i am making
this change I am bug hunting an elusive bug and I am trying to keep
things as much the same as they where from a time when things worked.
(This honestly won't do anything... but why not... If I knew where I
lost my keys, I wouldn't have lost keys ;-))