Connection Manager's Timer for re-pushing registration now put in conn mgr event queue;
To avoid threading issues we put calls that manipulate local state in
any way inside of the event queue for the component.  In this case the
quiesence re-push timer was peforming duties within its own thread.
This is deferred and run by the connection manager's event handling
thread.  This is the cleaner threadsafe thing to do.