Bug Fix - Because of initialization of ConnMgr before Registry...
Because The ConnMgr is instantiated before the Registry and it's
initialization process is kicked off first... part of it's start up is
to fire off a thread that sends off the local registry information to
the default peer.

The problem with that is that it happens before we even get our
registry written! Because the registry is instantiated after the
connection manager. :-(... This generated an NPE.  The strange issue
was that I was not able to get any stacktrace output that called out
the NPE from esgf-node1.  I have no idea why.  When the install was
done on the other test nodes the NPE was made clear and thus able to
find and fix the bug.

Whew, Now we can make a cut.
-gavin