Bug fix: Wrapped "poke", in the shards gleaner, in try catch...
For older nodes that have not updated their node manager, the port
value is not set and thus is written in the esgf_shards.xml with a
"null" where the port value should be.  When poke is run, null, is
clearly not an accpetable value and thus throws and exception.  The
catch for that exception is outside of the loop that builds the shards
list thus no more subsequent shards are recorded.

Now that this is wrapped directly, when this is encountered only the
trouble making shard is ignored and the rest of the shards poking and
writing can continue without a hitch.

This was sent in by Luca:
<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
<shards xmlns="http://www.esgf.org/whitelist">
    <value>adm07.cmcc.it:null/solr</value>
    <value>dev.esg.anl.gov:8983/solr</value>
    <value>esg-datanode.jpl.nasa.gov:8983/solr</value>
    <value>esg-vm-demo02.ccs.ornl.gov:8983/solr</value>
    <value>pcmdi11.llnl.gov:8983/solr</value>
    <value>pcmdi9.llnl.gov:null/solr</value>
</shards>

Notice that pcmdi9 and the first machine, Sandro's node, have "null"
for port values.