When pruning pull out the timestamp of node entry in question...
Use the node in question's time entry as the "value" in the removedMap.
This way we compare apples to apples. More specifically we compare
timestamps for the same node only.

With this fix... the timestamp of the same node is the ONLY timestamp
that is referenced when making decisions about what to do about that
node's representation in the merged registration.xml.

So the key change here is that when pruning, the timestamp set for
that action that is associated with the node, is the timestamp of that
node entry, NOT when the pruning occured (wall clock).  Upon an
encounter from another node that has the previously pruned node
present in their registration... the node's re-integration into the
current node's registration.xml is based on having a newer timestamp
for THAT node.  This should prevent the "bouncing" thus far seen.

A la vector clocks we only compare the clocks associated with a given
node.

-gavin