Added notion of - number of users that performed downloads to Metrics tag in registration.xml
We should be showing the number of unique users that have performed
downloads as part of the Metrics section of the registration.xml.
Previously the only value posted regarding users, was a count of the
registered users.

I felt this was limiting and incongruous with the larger notion of
"metrics".  Limiting because the only user metrics we were collecting
referred to the number of users that are registered on the node.  This
means that the node would have to be an IDP node to have that
information.  This is incongruous with the notion that download stats
are strictly under the Data node purview.  So right there we are
'crossing the streams'.  Also for completeness it makes more 'sense'
to give the notion of user in the context of downloading, specifically
the magnitude of the set of users that have been performing the
downloads.

The modifications in this commit would produce this bit of XML in the
registration.xml file:

 <Metrics>
     <DownloadedData count="32" size="28000" users="20"/>
     <RegisteredUsers count="4"/>
 </Metrics>

From the generated file containing:

DOWNLOADCOUNT=1,DOWNLOADSIZE=2,DOWNLOADUSERS=5,REGISTEREDUSERS=4

Where the last element is optional.
If it is "0" or null the resulting xml output will be.

 <Metrics>
     <DownloadedData count="32" size="28000" users="20"/>
 </Metrics>