WIP: Way of counting bytes sent through stream to end-user such that it can be recorded in access log
The main issue is that we need to get definitive stats on file
access/download.  Till now there has been a heuristic applied to the
only bit of data we really had which was the duration of download
time.  This, however, didn't give the logging fidelity we needed.

After a bit of digging, the least invasive mechanism was to augment
the AccessLoggingFilter so that it would pass data through a new
output stream.  This output stream would simply count the bits being
posted and, via a callback, send the number of total bytes moved into
the database.

What is missing is determining the bytes in the file in question.  And
also... to test if this thing works!!! :-)  It compiles, which usually
means it is all good (famous last words).