Re-work of how AccessLoggingFilter handles/resolves ingress url requests
We check a few patterns as urls are passed through the request.

(Still have to see how to write non-transformed paths to the path for
what this filter is fronting).

Only .nc file should be let in, by default.
If the *.nc file can be resolved via the mount points it will be.
If NOT then the path is passed directly.
The problem is that currently don't know the absolute path prefix to
the url file path.  Though there should be no .nc files that are not
mounted.